# public/api-docs/openapi.yaml
...
openapi: 3.1.0
info:
  title: GhostClass API
  version: 1.3.3
  description: |
    **GhostClass API** provides endpoints for managing attendance synchronization with EzyGo.
    
    ## Authentication
    
    The API supports two authentication methods:
    
    1. **Bearer Token (CRON_SECRET)**: For automated cron jobs
       - Add `Authorization: Bearer <CRON_SECRET>` header
       - Used by automated sync processes
    
    2. **Session Cookie (SupabaseAuth)**: For authenticated users
       - Automatically included when logged in through the web interface
       - Managed by Supabase Auth
    
    ## Rate Limiting
    
    All endpoints are rate-limited to prevent abuse:
    
    - `/api/auth/save-token` & `/api/cron/sync`: Configurable via environment variables
    - Contact Form (Server Action): Configurable via environment variables
    
    Rate limit information is returned in response headers:
    - `X-RateLimit-Limit`: Maximum requests allowed
    - `X-RateLimit-Remaining`: Remaining requests in current window
    - `X-RateLimit-Reset`: Unix timestamp when limit resets
    
    ## Base URLs
    
    - Production: Configured via `NEXT_PUBLIC_APP_URL`
    - Development: `http://localhost:3000`
    
    ## Error Handling
    
    All error responses follow a consistent format with appropriate HTTP status codes and descriptive messages.

  contact:
    name: GhostClass Support
    email: ${NEXT_PUBLIC_APP_EMAIL}
    url: ${NEXT_PUBLIC_GITHUB_URL}
  
  license:
...
servers:
...
paths:
  /api/auth/save-token:
    post:
      tags:
        - Authentication
      summary: Save EzyGo authentication token
      security: []  # No authentication required for this endpoint
      description: |
        Saves the EzyGo JWT authentication token and establishes a secure session for the user.
        
        **What it does:**
        - Validates the EzyGo token by verifying with EzyGo API
        - Creates or retrieves a Supabase authentication account (ghost login)
        - Encrypts and stores the token securely in the database
        - Establishes a session cookie for subsequent authenticated requests
        
        **Authentication:** No authentication required (this endpoint creates authentication)
        
        **Rate Limiting:** Configured via AUTH_RATE_LIMIT_REQUESTS and AUTH_RATE_LIMIT_WINDOW environment variables
        
        **Side Effects:**
        - Creates a new user account in Supabase Auth if not exists
        - Stores encrypted token in the database
        - Sets session cookies in the response
        
        **Use Cases:**
        - Initial user authentication from EzyGo
        - Token refresh when existing token expires
      
      operationId: saveAuthToken
      
      requestBody:
        required: true
        description: EzyGo authentication token
        content:
          application/json:
            schema:
              type: object
              required:
                - token
              properties:
                token:
                  type: string
                  minLength: 20
                  maxLength: 2000
                  description: Valid EzyGo JWT token obtained from EzyGo authentication
                  example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
            examples:
              validToken:
                summary: Valid EzyGo token
                description: Example of a valid JWT token from EzyGo
                value:
                  token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwidXNlcm5hbWUiOiJzdHVkZW50MTIzIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c"
      
      responses:
        '200':
          description: Token successfully saved and session established
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
              examples:
                success:
                  summary: Successful authentication
                  value:
                    success: true
        
        '400':
          description: Bad request - Missing or invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
              examples:
                missingToken:
                  summary: Token not provided
                  value:
                    message: "Missing credentials"
        
        '401':
          description: Unauthorized - Token verification failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                invalidToken:
                  summary: Invalid or expired token
                  value:
                    message: "Invalid or expired token"
                verificationFailed:
                  summary: Could not verify user identity
                  value:
                    message: "Could not verify user identity"
        
        '429':
          description: Rate limit exceeded
          headers:
            X-RateLimit-Limit:
              schema:
                type: integer
              description: Maximum number of requests allowed in the time window (from AUTH_RATE_LIMIT_REQUESTS)
              example: 5
            X-RateLimit-Remaining:
              schema:
                type: integer
              description: Number of requests remaining in current window
              example: 0
            X-RateLimit-Reset:
              schema:
                type: integer
              description: Unix timestamp when the rate limit resets
              example: 1706108400
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RateLimitError'
              examples:
                tooManyRequests:
                  summary: Rate limit exceeded
                  value:
                    error: "Too many requests. Slow down!"
                    retryAfter: 900
        
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                serverError:
                  summary: Failed to establish session
                  value:
                    message: "Failed to establish secure session"
...
components:
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message describing what went wrong
        message:
          type: string
          description: Detailed error message
      example:
        error: "Internal Server Error"
    ValidationError:
      type: object
      properties:
        message:
          type: string
          description: Validation error message
        errors:
          type: array
          description: List of validation errors (if multiple fields failed)
          items:
            type: object
            properties:
              field:
                type: string
                description: Field that failed validation
              message:
                type: string
                description: Validation error message for this field
      example:
        message: "Missing credentials"
    RateLimitError:
      type: object
      properties:
        error:
          type: string
          description: Rate limit error message
          example: "Too many requests. Slow down!"
        retryAfter:
          type: integer
          description: Unix timestamp (in seconds) when the rate limit resets. Matches the `X-RateLimit-Reset` header.
          example: 1735689600
      required:
        - error
        - retryAfter
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Bearer token authentication using CRON_SECRET environment variable.
        Used for automated cron jobs.
        
        Example: `Authorization: Bearer <CRON_SECRET>`
    
    SupabaseAuth:
      type: apiKey
      in: cookie
      name: sb-*
      description: |
        Session cookie authentication managed by Supabase Auth.
        Automatically included when users are logged in through the web interface.
        Cookie names typically start with 'sb-' prefix.
...